Conflicts:
	modules/partners/kimportal/src/test/java/KIMPortalPartnerRecommenderTest.java
	pom.xml

+++ modules/recommender/federated-recommender/src/main/java/eu/eexcess/federatedrecommender/dataformats/ResultSimilarity.java
@@ -87,7 +87,7 @@ public class ResultSimilarity extends Result implements Serializable {
             valueCount += 3.0;
         }
 
-        if (valueCount == 0.0)
+        if (valueCount.intValue()==0)
             return 0.0;
         return returnValue / valueCount;
     }
