+++ modules/recommender/federated-recommender/src/main/java/eu/eexcess/federatedrecommender/domaindetection/wordnet/WordnetDomainsDetector.java
@@ -57,9 +57,9 @@ import at.knowcenter.ie.opennlp.TokenAnnotator;
 import at.knowcenter.ie.postags.UnifiedTag;
 import at.knowcenter.util.term.TermSet;
 import at.knowcenter.util.term.TypedTerm;
-import eu.eexcess.federatedrecommender.domaindetection.Domain;
-import eu.eexcess.federatedrecommender.domaindetection.DomainDetector;
-import eu.eexcess.federatedrecommender.domaindetection.DomainDetectorException;
+import eu.eexcess.federatedrecommender.domaindetection.probing.Domain;
+import eu.eexcess.federatedrecommender.domaindetection.probing.DomainDetector;
+import eu.eexcess.federatedrecommender.domaindetection.probing.DomainDetectorException;
 
 /**
  * Domain detector based on the wordnet domains dataset.
@@ -171,7 +171,7 @@ public class WordnetDomainsDetector extends DomainDetector {
 				Set<String> allDomains = new HashSet<String>();
 				List<Synset> senses = randomIndexWord.getSenses();
 				for (Synset synset : senses) {
-					String synsetId = String.format("%08d-%s", synset.getKey(), synset.getPOS().getKey());
+					String synsetId = String.format("%08d-%s", (long) synset.getKey(), synset.getPOS().getKey());
 					Set<DomainAssignment> domains = synsetToDomains.get(synsetId);
 					if (domains != null && !domains.isEmpty()) {
 						for (DomainAssignment domain : domains) {
@@ -395,7 +395,7 @@ public class WordnetDomainsDetector extends DomainDetector {
 	 * @param synset
 	 */
 	protected void collectDomains(Synset synset, Map<Synset, Set<DomainAssignment>> synsetToAssignments) {
-		String synsetId = String.format("%08d-%s", synset.getKey(), synset.getPOS().getKey());
+		String synsetId = String.format("%08d-%s", (long)synset.getKey(), synset.getPOS().getKey());
 		Set<DomainAssignment> domains = synsetToDomains.get(synsetId);
 		if (domains != null && !domains.isEmpty()) {
 			// System.out.println("  "+synset.getKey()+", "+synset+", "+domains);
